scrape_cinemas/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
flask==3.0.3
flask-swagger-ui==4.11.1
python-dotenv==1.0.1
azure-cosmos==4.7.0
flask-cors==5.0.0